Added TD(0) agent and changed other stuff appropriately
Changed value functions to be more like action-value functions, where we only store the value of game states AFTER we make a move. In this way the game states encode actions our agent may take. Agent has no value for states where it is the agent's turn.

Changed Agent class structure to be slightly more modular. There is an Agent base class that can play moves but cannot learn. MonteCarloAgent and TDAgent extend Agent and implement a function to learn.